Shader compila

+++ src/effects/outline/componets.rs
@@ -3,5 +3,5 @@ use bevy::{color::Color, ecs::component::Component};
 #[derive(Component)]
 pub struct SpriteOutline {
     pub color: Color,
-    pub thickness: u8,
+    pub thickness: f32,
 }
